Refactor "Scrobble" class into it's own file.
